Restructure into a plugin.
