+++ wordpress.org/public_html/wp-content/mu-plugins/pub/wporg-seo/redirects.php
@@ -1,0 +1,25 @@
+<?php
+namespace WordPressdotorg\SEO\Redirects;
+
+/**
+ * Custom Canonical redirect for Facebook and Twitter referrers.
+ */
+function facebook_twitter_referers() {
+	// Only run on pages with canonical enabled.
+	if ( ! has_action( 'template_redirect', 'redirect_canonical' ) ) {
+		return;
+	}
+
+	$url = false;
+	if ( isset( $_GET['fbclid'] ) ) {
+		$url = remove_query_arg( 'fbclid' ) . '#utm_medium=referral&utm_source=facebook.com&utm_content=social';
+	} elseif ( isset( $_GET['__twitter_impression'] ) ) {
+		$url = remove_query_arg( '__twitter_impression' ) . '#utm_medium=referral&utm_source=twitter.com&utm_content=social';
+	}
+
+	if ( $url ) {
+		wp_safe_redirect( $url, 301 );
+		exit;
+	}
+}
+add_action( 'template_redirect', __NAMESPACE__ . '\facebook_twitter_referers', 9 ); // Before redirect_canonical();
