+++ wordpress.org/public_html/wp-content/mu-plugins/pub/wporg-seo/wporg-seo.php
@@ -1,0 +1,9 @@
+<?php
+namespace WordPressdotorg\SEO;
+/**
+ * Plugin Name: WordPress.org Generic SEO fixes.
+ */
+
+include __DIR__ . '/robots.php';
+include __DIR__ . '/redirects.php';
+include __DIR__ . '/canonical.php';
